seed/requirements.txt
pyyaml
matplotlib
numpy
scipy
flake8